Add AbstractAmplitude class
